requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
passlib[bcrypt]

fpdf2
pymysql

